+++ spring-testcontainer/spring-testcontainer-electrum-daemon-starter/src/main/java/org/tbk/spring/testcontainer/electrumd/config/ElectrumDaemonContainerAutoConfiguration.java
@@ -45,6 +45,7 @@ public class ElectrumDaemonContainerAutoConfiguration {
         return ElectrumDaemonContainerConfig.builder()
                 .environment(properties.getEnvironmentWithDefaults())
                 .defaultWallet(properties.getDefaultWallet().orElse(null))
+                .wallets(properties.getWallets())
                 .build();
     }